+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0
+- Store state in a DbSession
+- Fix UTF-8
+- Fire events when nodes are created, updated or deleted
+
 ## 0.1.1
 - Polymorphism in OGM (Support returning and saving subclasses of T in repository)